[NET] 2023-04-10 #01: - separated files for actor, critic and common network features
