+++ networks/common.py
@@ -1,0 +1,77 @@
+import ptan
+import torch
+import math
+import numpy as np
+
+from config import *
+
+#**************************** AGENT MODEL ******************************
+
+class Agent(ptan.agent.BaseAgent):
+    def __init__(self, net, device="cpu"):
+        self.net = net
+        self.device = device
+
+    def __call__(self, states, agent_states):
+        states_v = ptan.agent.float32_preprocessor(states)
+        states_v = states_v.to(self.device)
+
+        mu_v = self.net(states_v)
+        mu = mu_v.data.cpu().numpy()
+        logstd = self.net.logstd.data.cpu().numpy()
+        rnd = np.random.normal(size=logstd.shape)
+        actions = mu + np.exp(logstd) * rnd
+        actions = np.clip(actions, -1, 1)
+        return actions, agent_states
+    
+
+#**************************** FUNCTIONS ********************************
+
+def test_net(net, env, device, count=10):
+    rewards = 0.0
+    steps = 0
+    for _ in range(count):
+        obs = env.reset()
+        while True:
+            obs_v = ptan.agent.float32_preprocessor([obs]).to(device)
+            mu_v = net(obs_v)[0]
+            action = mu_v.squeeze(dim=0).data.cpu().numpy()
+            action = np.clip(action, -1, 1)
+            obs, reward, done, _ = env.step(action)
+            rewards += reward
+            steps += 1
+            if done:
+                break
+    return rewards / count, steps / count
+
+
+def calc_logprob(mu_v, logstd_v, actions_v):
+    p1 = - ((mu_v - actions_v) ** 2) / (2*torch.exp(logstd_v).clamp(min=1e-3))
+    p2 = - torch.log(torch.sqrt(2 * math.pi * torch.exp(logstd_v)))
+    return p1 + p2
+
+
+
+def calc_adv_ref(trajectory, net_crt, states_v, device):
+    
+    values_v = net_crt(states_v)
+    values = values_v.squeeze().data.cpu().numpy()
+
+    last_gae = 0.0
+    result_adv = []
+    result_ref = []
+    for val, next_val, (exp,) in zip(reversed(values[:-1]),
+                                     reversed(values[1:]),
+                                     reversed(trajectory[:-1])):
+        if exp.done:
+            delta = exp.reward - val
+            last_gae = delta
+        else:
+            delta = exp.reward + GAMMA * next_val - val
+            last_gae = delta + GAMMA * GAE_LAMBDA * last_gae
+        result_adv.append(last_gae)
+        result_ref.append(last_gae + val)
+
+    adv_v = torch.FloatTensor(list(reversed(result_adv)))
+    ref_v = torch.FloatTensor(list(reversed(result_ref)))
+    return adv_v.to(device), ref_v.to(device)
